% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwolf_sim.R
\name{fwer_sim}
\alias{fwer_sim}
\title{Simulate data as in Clarke, Romano & Wolf (2019) to simulate family wise
error rates (FWERs)}
\usage{
fwer_sim(rho, N, s, B)
}
\arguments{
\item{rho}{The correlation between the outcome variables}

\item{N}{The number of observations}

\item{s}{The number of dependent variables}

\item{B}{The number of bootstrap draws employed in the 'rwolf()' and
'wyoung()' procedures}
}
\value{
A 'data.frame' containing unadjusted p-values & p-values adjusted using the
methods by Holm and Romano & Wolf (2005)
}
\description{
Simulate data as in Clarke, Romano & Wolf (2019) to simulate family wise
error rates (FWERs)
}
