% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rwolf_pval.R
\name{get_rwolf_pval}
\alias{get_rwolf_pval}
\title{compute Romano-Wolf adjusted p-values based 
on bootstrapped t-statistics}
\usage{
get_rwolf_pval(t_stats, boot_t_stats)
}
\arguments{
\item{t_stats}{A vector of length S - where S is the number of 
tested hypotheses - containing the original, 
non-bootstrappe t-statisics}

\item{boot_t_stats}{A (B x S) matrix containing the 
bootstrapped t-statistics}
}
\value{
A vector of Romano-Wolf corrected p-values
}
\description{
compute Romano-Wolf adjusted p-values based 
on bootstrapped t-statistics
}
