% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwolf_sim.R
\name{run_fwer_sim}
\alias{run_fwer_sim}
\title{Run a MC simulation study on family-wise error rates (FWERs)
for the Holm, Westfall & Young and Romano & Wolf Methods multiple
hypothesis adjustment methods}
\usage{
run_fwer_sim(
  n_sims = 100,
  rho = c(0, 0.25, 0.5, 0.75),
  seed = 114411,
  B = 499,
  N = 1000,
  s = 6
)
}
\arguments{
\item{n_sims}{The number of Monte Carlo iterations. 1000 by default.}

\item{rho}{The correlation between the outcome variables.}

\item{seed}{A random seed.}

\item{B}{The number of bootstrap draws emplozed in the 'rwolf()' and
'wyoung' procedures}

\item{N}{The number of observations}

\item{s}{The number of dependent variables}
}
\description{
In general, this is a wrapper around `fwer_sim()`. See `?fwer_sim()`
for more information
}
\examples{
\dontrun{

res <- run_fwer_sim(
  seed = 123,
  n_sims = 1000,
  B = 4999,
  N = 1000,
  s = 10
)
res

}
}
