% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwolf.R
\name{rwolf}
\alias{rwolf}
\title{Romano-Wolf multiple hypotheses adjusted p-values}
\usage{
rwolf(models, param, B, seed = NULL, package = "fwildclusterboot", ...)
}
\arguments{
\item{models}{An object of type fixest_multi}

\item{param}{The regression param to be tested}

\item{B}{The number of bootstrap iterations}

\item{seed}{Integer. Sets the random seed. NULL by default.}

\item{package}{Should the wild cluster bootstrap run via fwildclusterboot or wildboottestjlr? fwildclusterboot by default}

\item{...}{additional function values passed to the bootstrap function.}
}
\description{
Function implements the Romano-Wolf multiple hypthesis correction procedure for objects of type fixest_multi (fixest_multi are objects created by `fixest::feols()` that use `feols()` multiple-estimation interface). 
Currently, the command is restricted to two-sided hypotheses and oneway clustered standard errors. For the wild cluster bootstrap, 
the null is always imposed.
}
\examples{
 
library(fixest)
library(rwolf)

set.seed(12345)

N <- 1000
X1 <- rnorm(N)
Y1 <- 1 + 1 * X1 + rnorm(N)
Y2 <- 1 + 0.01 * X1 + rnorm(N)
Y3 <- 1 + 0.01 * X1 + rnorm(N)
Y4 <- 1 + 0.01 * X1 + rnorm(N)

B <- 100
# intra-cluster correlation of 0 for all clusters
cluster <- rep(1:50, N / 50)

data <- data.frame(Y1 = Y1, 
                   Y2 = Y2, 
                   Y3 = Y3, 
                   Y4 = Y4,
                   X1 = X1, 
                   cluster = cluster)

res <- feols(c(Y1, Y2, Y3, Y4) ~ X1, data = data, cluster = ~ cluster)
res_rwolf <- rwolf(models = res, param = "X1", B = B)
summary(res_rwolf)

}
\references{
Clarke, Romano & Wolf (2019), STATA Journal. IZA working paper: https://ftp.iza.org/dp12845.pdf
}
