% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwolf.R
\name{summary.rwolf}
\alias{summary.rwolf}
\title{Summary method for objects of type rwolf}
\usage{
\method{summary}{rwolf}(object, digits, ...)
}
\arguments{
\item{object}{An object of type rwolf}

\item{digits}{Rounding of digits}

\item{...}{misc. function arguments}
}
\description{
Summary method for objects of type rwolf
}
