% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwolf.R
\name{rwolf}
\alias{rwolf}
\title{Function implements the Romano-Wolf multiple hypthesis correction procedure
for objects of type fixest_multi}
\usage{
rwolf(models, parameter, B, seed = NULL)
}
\arguments{
\item{models}{An object of type fixest_multo}

\item{parameter}{The regression parameter to be tested}

\item{B}{The number of bootstrap iterations}

\item{seed}{Integer. Sets the random seed}
}
\description{
Function implements the Romano-Wolf multiple hypthesis correction procedure
for objects of type fixest_multi
}
